Moved to frontend and added additional variables to the garage class.
